Upgpkg: 1.9-1
Add suspending capability for the spigot server through the control
script and an additional screen session which listens on the same port
with netcat.
        * Update .conf accordingly
        * IDLE_SERVER can be either true or false
        * The default is no idle server
        * Choose an available netcat
	* flavor automatically
Move server root to /srv/craftbukkit .
Variable renaming in .install.
Notify the user of kept directories on post_remove and alter ownership
to root before removing the game user.
Print error messages to stderr.