Upgpkg: 1.9-1
Fix issues with how the idle_server is handled.
On server stop, return to the prompt as soon as the server has
completely shut down, not as soon as the shutdown command has been sent.
Thereby fixing the restart capability.
Introduce sanity check whether the servers is really down in idle_daemon
and server_start as well.

Signed-off-by: Gordian Edenhofer <gordian.edenhofer@gmail.com>