Upgpkg: 1.9.2-1
The list command prints a line containing the usernames after the last
occurrence of ": " and since playernames may not contain this string the
clean player-list can be easily retrieved through the use of sed. This
surpasses the usefulness of the player_delimiter.
Increase waiting period between executing the list game-command and
scanning the logs from 0.3 seconds to 0.6 seconds.
Tweak tail -f command in server_command.
Use ${_game} variable aggressively in PKGBUILD.