Merge letsencrypt-git into certbot-git
The package has been renamed a while ago and both versions were kept
initially for a period of time to ease the switch. Since the new
package is, as of now, more than a month old it is probably safe to
assume that most users have already updated their system or will not do
so in the foreseeable future.