upgpkg: minecraft-server 1.11.2-2
Change the idle_server naming schema in the script.

Previously screen was used to select a session with a specific well
defined name however apparently screen selects anything starting with
the specified pattern. This let to some confusion between the server and
the idle_server. To circumvent this behaviour simply change the naming
schema of the idle_server. THIS IS NOT A FIX! - part 2